Resolve without the versions plugin
